requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-genai
python-dotenv
pymupdf
python-docx